+++ api/tests/routes/public/individual_offers/v1/get_events_test.py
@@ -1,0 +1,49 @@
+import pytest
+
+from pcapi.core import testing
+from pcapi.core.offerers import factories as offerers_factories
+from pcapi.core.offers import factories as offers_factories
+
+from . import utils
+
+
+@pytest.mark.usefixtures("db_session")
+class GetEventsTest:
+    ENDPOINT_URL = "http://localhost/public/offers/v1/events"
+
+    def test_get_first_page(self, client):
+        venue, _ = utils.create_offerer_provider_linked_to_venue()
+        offers = offers_factories.EventOfferFactory.create_batch(12, venue=venue)
+        offers_factories.ThingOfferFactory.create_batch(3, venue=venue)  # not returned
+
+        with testing.assert_no_duplicated_queries():
+            response = client.with_explicit_token(offerers_factories.DEFAULT_CLEAR_API_KEY).get(
+                f"/public/offers/v1/events?limit=5&venueId={venue.id}"
+            )
+
+        assert response.status_code == 200
+        assert response.json["pagination"] == {
+            "currentPage": 1,
+            "itemsCount": 5,
+            "itemsTotal": 12,
+            "lastPage": 3,
+            "limitPerPage": 5,
+            "pagesLinks": {
+                "current": f"{self.ENDPOINT_URL}?venueId={venue.id}&page=1&limit=5",
+                "first": f"{self.ENDPOINT_URL}?venueId={venue.id}&page=1&limit=5",
+                "last": f"{self.ENDPOINT_URL}?venueId={venue.id}&page=3&limit=5",
+                "next": f"{self.ENDPOINT_URL}?venueId={venue.id}&page=2&limit=5",
+                "previous": None,
+            },
+        }
+        assert [event["id"] for event in response.json["events"]] == [offer.id for offer in offers[0:5]]
+
+    def test_404_when_venue_id_not_tied_to_api_key(self, client):
+        offerers_factories.ApiKeyFactory()
+        unrelated_venue = offerers_factories.VenueFactory()
+
+        response = client.with_explicit_token(offerers_factories.DEFAULT_CLEAR_API_KEY).get(
+            f"/public/offers/v1/events?venueId={unrelated_venue.id}"
+        )
+
+        assert response.status_code == 404
